Phase 3: Decompose QueryEngine into focused modules
- Extract type alias resolution into clang_index_mcp/_search/type_alias_resolver.py.
- Extract file-scoped symbol lookup into clang_index_mcp/_search/file_symbol_finder.py
  (find_in_file, get_files_containing_symbol, path suggestion helpers).
- Extract template inheritance analysis into clang_index_mcp/_search/template_analyzer.py
  (template arg parsing, indirect inheritance, get_derived_classes).
- Extract hierarchy traversal into clang_index_mcp/_search/hierarchy_analyzer.py
  (BFS traversal, get_class_hierarchy).
- Keep thin public wrappers in QueryEngine so CppAnalyzer callers remain unchanged.
- Full test suite passes: 1432 passed, 17 skipped, 4 xpassed.
